server/requirements.txt
Django>=2.1,<2.2
Pygments==2.2.0
backcall==0.1.0
coverage
decorator==4.3.0
django-cors-headers==2.4.0
django-filter==2.2.0
django-silk
djangorestframework-camel-case==0.2.0
djangorestframework>=3.9.1,<3.10
ipython-genutils==0.2.0
ipython==6.5.0
jedi==0.12.1
parameterized==0.7.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2-binary==2.7.4
ptyprocess==0.6.0
pytest-django==3.4.3
pytest==3.9.3
pytz==2018.5
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
uwsgi
wcwidth==0.1.7
xmltodict==0.11.0

drf-yasg[validation]
packaging
